refactor: code for MenghapusPengajuanSurat

+++ situ/src/main/java/apap/tugasakhir/situ/service/PengajuanSuratService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface PengajuanSuratService {
     List<PengajuanSuratModel> getListPengajuanSurat();
     PengajuanSuratModel addPengajuanSurat(PengajuanSuratModel pengajuanSurat);
-    String deletePengajuanSurat(Long id);
+    String deletePengajuanSurat(Integer id);
     PengajuanSuratModel ubahPengajuanSurat(PengajuanSuratModel pengajuanSurat);
 }
